requirements.txt
certifi==2023.7.22
dataclasses==0.6
dnspython==2.4.2
from-root==1.3.0
joblib==1.3.2
numpy==1.24.4
pandas==2.0.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.10.1
-e git+https://github.com/sachint97/sensor-fault-detection.git@ba8b484fd735f5e3482755ff75e85541504c7f9d#egg=sensor_fault_detection
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
